args: support plugin-specific JSON config
Currently the `tc-redirect-tap` plugin only supports the arguments it receives
via the `skel.CmdArgs` arguments, which in turn come the `CNI_ARGS` environment
variable. To allow driving the configuration purely from files and bring the
plugin configuration more in line with the default plugins, add support for
plugin-specific JSON configuration.

Signed-off-by: Tim Gross <tim@0x74696d.com>