Return correct CNI version string
Prior to this commit, we were using the output of a result to return a
CNI version. This would always be the latest supported version returned
by the CNI library. This meant that, regardless of the CNI version used
in a user's conflist, it would always report back that the latest
version was being used, which is both incorrect behavior and breaks
workflows where the latest version was not recognized as a valid
version.

This change fixes this behavior and returns the same CNI version
specified in the conflist, which mirrors the behavior of other CNI
plugins.

Signed-off-by: David Son <davbson@amazon.com>